Decrement numberOfPreviewAssets by 1 for endIndex
The numberOfPreviewAssets prop should be decremented by 1 when setting the endIndex in order to fetch the proper amount of previewAssets.  Currently, if the previewAssets prop is set to 1, you will receive two assets.  Decrementing the numberOfPreviewAssets by 1 when setting the endIndex seems to be a sufficient place to implement this change in order to achieve desired functionality of previewAssets.  I have tested this patch and am now receiving the correct amount of previewAssets.