Updated indexRangeLength and assetDisplayBottomUp
The indexSet was not calculating the correct range length when sorting from oldest to newest when starting at index 0.  Since the range was being calculated by endIndex - startIndex, the range was off by 1 since the actual range of assets for 0-2 is 3 and not 2. I created an if to account for this if the originalStartIndex and startIndex are both 0.

I also adjusted the way the assets are displayed from the bottom to top of page depending on whether assets should be loaded from start to end.

I've tested against the test scenarios and all is working.